requirements.txt
google-generativeai
python-dotenv
openai
# Force rebuild for Gemini switch
